requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
aiofiles
numpy
cairosvg
Pillow
